tarea: usar barrios del registro
