+++ app/inscripcion/templates/inscripcion/nueva.html
@@ -33,7 +33,7 @@
     import Autocomplete from '/static/js/autocomplete.js'
     new Autocomplete(document.getElementById('id_barrio_tutor'), {
       liveServer: true,
-      server: '/inscripcion/barrios/',
+      server: '/registro/barrios/',
       serverParams: { q: 'typed' },
       valueField: 'id',
       labelField: 'title',
