+++ app/inscripcion/urls.py
@@ -6,5 +6,4 @@ app_name = "inscripcion"
 urlpatterns = [
     path("<int:campana_id>/", views.index, name="index"),
     path("crear/<int:campana_id>/", views.crear, name="create"),
-    path("barrios/", views.obtener_barrios, name="barrios"),
 ]
