+++ app/registro/templates/registro/nuevo.html
@@ -40,7 +40,7 @@
     })
     new Autocomplete(document.getElementById('id_barrio_tutor'), {
       liveServer: true,
-      server: '/inscripcion/barrios/',
+      server: '/registro/barrios/',
       serverParams: { q: 'typed' },
       valueField: 'id',
       labelField: 'title',
