+++ app/registro/urls.py
@@ -14,4 +14,5 @@ urlpatterns = [
     path("vista_veterinarios/<int:campana_id>", views.RegistradoListView.as_view(), name="vista_veterinarios"),
     path("registro/<int:registro_id>/generar_pdf/", views.generar_pdf, name="generar_pdf"),
     path("razas/", views.obtener_razas, name="razas"),
+    path("barrios/", views.obtener_barrios, name="barrios"),
 ]
